+++ packages/kununu-footer/index.module.scss
