+++ packages/kununu-header/package.json
@@ -19,7 +19,7 @@
   "scripts": {
     "prepare": "npm run dist",
     "babel": "babel $(find ./ -name '*.jsx' ! -name '*.spec.jsx' ! -ipath '*node_modules*') --out-dir ./dist --root-mode upward",
-    "sass": "node-sass ./index.scss --output ./dist --importer node_modules/node-sass-module-importer",
+    "sass": "node-sass ./index.module.scss --output ./dist --importer node_modules/node-sass-module-importer",
     "clean": "rimraf ./dist",
     "dist": "npm run clean && cross-env NODE_ENV=production npm run babel && npm run sass"
   }
